introduce debug_assert!s in endorelation-operations
All endorelation-operations require the underlying relation to
be homogeneous (i.e. an endo-relation). Verifying this for every
operation would however incur significant overhead and slow down
the library even when the user has already made sure that a given
relation is indeed homogeneous.
debug_assert!s serve as a compromise in this situation: they are
removed from release targets and thus do not slow down the library,
while the checks are still performed in development.